refactored userSkillId field
1. Needed to NOT pass in the field userSkillId when it is randomly generated by the transformSkillsAndRatings function. This allows the mutation to correctly save the other info to the DB and randomly generate it's own userSkillId
