+++ pages/api/skillRatingsFunctions.ts
@@ -1,4 +1,5 @@
 import { SkillsAndRatings } from "../../graphql/fetchSkillsAndRatings";
+import { User } from "../../graphql/fetchUserProfile";
 import { UserSkillsRatings } from "../../graphql/fetchUserSkillsRatings";
 import { SkillRatingsRow } from "../../redux/skillRatingsSlice";
 
@@ -24,10 +25,32 @@ import { SkillRatingsRow } from "../../redux/skillRatingsSlice";
         skillId: row.id,
         skillName: row.name,
         unitName: row.intro_course_unit.title,
-        studentRating: row.intro_course_skills_users_aggregate.nodes.length > 0 && row.intro_course_skills_users_aggregate.nodes[0].studentRating
+        studentRating: row.intro_course_skills_users_aggregate.nodes.length > 0 
         ? parseInt(row.intro_course_skills_users_aggregate.nodes[0].studentRating)
         : 0,
-        userSkillId: row.intro_course_skills_users[0]?.id,
+        userSkillId: row.intro_course_skills_users[0]?.id ? row.intro_course_skills_users[0].id : Math.random(),
       };
     });
+  };
+
+  export const transformSkillRatingForDB = (skillRatings: SkillRatingsRow[], user:User) => {
+    // map from redux type to write back to DB
+    const transformedOutput = skillRatings.map((row) => {
+      if (typeof row.userSkillId == "number") {
+        return {
+          userId: user.uid,
+          skillId: row.skillId,
+          studentRating: row.studentRating,
+        };
+      } else {
+        return {
+          userId: user.uid,
+          id: row.userSkillId,
+          skillId: row.skillId,
+          studentRating: row.studentRating,
+        };
+      }
+    });
+
+    return transformedOutput;
   };
